refactor(main): grouping tasks related files Signed-off-by: axel7083 <42176370+axel7083@users.noreply.github.com>
