+++ packages/main/src/plugin/extension-loader.spec.ts
@@ -57,13 +57,13 @@ import type { InputQuickPickRegistry } from './input-quickpick/input-quickpick-r
 import type { KubernetesClient } from './kubernetes-client.js';
 import type { MenuRegistry } from './menu-registry.js';
 import type { MessageBox } from './message-box.js';
-import type { NotificationRegistry } from './notification-registry.js';
 import type { OnboardingRegistry } from './onboarding-registry.js';
-import type { ProgressImpl } from './progress-impl.js';
 import type { ProviderRegistry } from './provider-registry.js';
 import type { Proxy } from './proxy.js';
 import type { ExtensionSecretStorage, SafeStorageRegistry } from './safe-storage/safe-storage-registry.js';
 import type { StatusBarRegistry } from './statusbar/statusbar-registry.js';
+import type { NotificationRegistry } from './tasks/notification-registry.js';
+import type { ProgressImpl } from './tasks/progress-impl.js';
 import type { Telemetry } from './telemetry/telemetry.js';
 import type { TrayMenuRegistry } from './tray-menu-registry.js';
 import type { IDisposable } from './types/disposable.js';
