+++ packages/main/src/plugin/extension-loader.ts
@@ -59,10 +59,7 @@ import { InputBoxValidationSeverity, QuickPickItemKind } from './input-quickpick
 import type { KubernetesClient } from './kubernetes-client.js';
 import type { MessageBox } from './message-box.js';
 import { ModuleLoader } from './module-loader.js';
-import type { NotificationRegistry } from './notification-registry.js';
 import type { OnboardingRegistry } from './onboarding-registry.js';
-import type { ProgressImpl } from './progress-impl.js';
-import { ProgressLocation } from './progress-impl.js';
 import type { ProviderRegistry } from './provider-registry.js';
 import type { Proxy } from './proxy.js';
 import { createHttpPatchedModules } from './proxy-resolver.js';
@@ -74,6 +71,9 @@ import {
   StatusBarItemImpl,
 } from './statusbar/statusbar-item.js';
 import type { StatusBarRegistry } from './statusbar/statusbar-registry.js';
+import type { NotificationRegistry } from './tasks/notification-registry.js';
+import type { ProgressImpl } from './tasks/progress-impl.js';
+import { ProgressLocation } from './tasks/progress-impl.js';
 import type { Telemetry } from './telemetry/telemetry.js';
 import type { TrayMenuRegistry } from './tray-menu-registry.js';
 import type { IDisposable } from './types/disposable.js';
