+++ packages/main/src/plugin/tasks/progress-impl.spec.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2023 Red Hat, Inc.
+ * Copyright (C) 2023-2024 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import { beforeEach, expect, test, vi } from 'vitest';
 import type { CommandRegistry } from '/@/plugin/command-registry.js';
 import type { StatusBarRegistry } from '/@/plugin/statusbar/statusbar-registry.js';
 
-import type { ApiSenderType } from './api.js';
+import type { ApiSenderType } from '../api.js';
 import { ProgressImpl, ProgressLocation } from './progress-impl.js';
 import { TaskManager } from './task-manager.js';
