+++ packages/main/src/plugin/tasks/progress-impl.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2022 Red Hat, Inc.
+ * Copyright (C) 2022-2024 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,10 +17,9 @@
  ***********************************************************************/
 import type * as extensionApi from '@podman-desktop/api';
 
-import type { TaskManager } from '/@/plugin/task-manager.js';
-
-import { findWindow } from '../util.js';
-import { CancellationTokenImpl } from './cancellation-token.js';
+import { findWindow } from '../../util.js';
+import { CancellationTokenImpl } from '../cancellation-token.js';
+import type { TaskManager } from './task-manager.js';
 
 export enum ProgressLocation {
   /**
